+++ controller/controllerList.js
@@ -11,6 +11,16 @@ async function addBookToClub(req, res) {
 
     try {
 
+        const idUser = req.user.idusers;
+        const checkUser = `SELECT * FROM users WHERE idusers = ?`
+        const [userExists] = await connection.query(checkUser, idUser);
+        
+        if (userExists[0].idusers !== req.user.idusers) {
+            return res.status(403).send({
+                message: "Você não tem permissão para criar reviews"
+            })
+        }
+
         const insert = `INSERT INTO book_club_book (book_id, book_club_id) VALUES ((SELECT idbooks FROM books WHERE titulo = ?), (SELECT idbookclub FROM book_club WHERE nomebookclub = ?));`
         const [result] = await connection.query(insert, [titulo, nomebookclub]);
         const newList = {
@@ -41,6 +51,17 @@ async function readBookInClub(req, res) {
     }
 
     try {
+
+        const idUser = req.user.idusers;
+        const checkUser = `SELECT * FROM users WHERE idusers = ?`
+        const [userExists] = await connection.query(checkUser, idUser);
+        
+        if (userExists[0].idusers !== req.user.idusers) {
+            return res.status(403).send({
+                message: "Você não tem permissão para criar reviews"
+            })
+        }
+
         // devido a tabela associativa a query precisa de um ou mais inner join para achar o id correto de acordo com o nosso where
         const selectByClub = `SELECT titulo, nomebookclub FROM book_club INNER JOIN book_club_book ON book_club.idbookclub = book_club_book.book_club_id INNER JOIN books ON book_club_book.book_id = books.idbooks WHERE nomebookclub = ?`;
         const books = await connection.query(selectByClub, nomebookclub);
@@ -68,6 +89,17 @@ async function readClubWithBook(req, res) {
     }
 
     try {
+
+        const idUser = req.user.idusers;
+        const checkUser = `SELECT * FROM users WHERE idusers = ?`
+        const [userExists] = await connection.query(checkUser, idUser);
+        
+        if (userExists[0].idusers !== req.user.idusers) {
+            return res.status(403).send({
+                message: "Você não tem permissão para criar reviews"
+            })
+        }
+
         // mesma logica da query acima
         const selectByBook = `SELECT nomebookclub FROM book_club INNER JOIN book_club_book ON book_club.idbookclub = book_club_book.book_club_id INNER JOIN books ON book_club_book.book_id = books.idbooks WHERE titulo = ?`;
         const books = await connection.query(selectByBook, titulo);
@@ -87,6 +119,16 @@ async function readClubWithBook(req, res) {
 async function readAllBooksClubs(req, res) {
 
     try {
+
+        const idUser = req.user.idusers;
+        const checkUser = `SELECT * FROM users WHERE idusers = ?`
+        const [userExists] = await connection.query(checkUser, idUser);
+        
+        if (userExists[0].idusers !== req.user.idusers) {
+            return res.status(403).send({
+                message: "Você não tem permissão para criar reviews"
+            })
+        }
 
         const select = `SELECT nomebookclub, titulo FROM book_club INNER JOIN book_club_book ON book_club.idbookclub = book_club_book.book_club_id INNER JOIN books ON book_club_book.book_id = books.idbooks`;
         const clubsbooks = await connection.query(select);
@@ -113,6 +155,16 @@ async function updateBookInClub(req, res) {
 
     try {
 
+        const idUser = req.user.idusers;
+        const checkUser = `SELECT * FROM users WHERE idusers = ?`
+        const [userExists] = await connection.query(checkUser, idUser);
+        
+        if (userExists[0].idusers !== req.user.idusers) {
+            return res.status(403).send({
+                message: "Você não tem permissão para criar reviews"
+            })
+        }
+
         const updateByName = `UPDATE book_club_book INNER JOIN books ON book_club_book.book_id = books.idbooks INNER JOIN book_club ON book_club_book.book_club_id = book_club.idbookclub SET book_id = (SELECT idbooks FROM books WHERE titulo = ?) WHERE nomebookclub = ?`;
         const books = await connection.query(updateByName, [titulonovo, nomebookclub]);
         connection.release();
@@ -138,6 +190,16 @@ async function updateClubInList(req, res) {
 
     try {
 
+        const idUser = req.user.idusers;
+        const checkUser = `SELECT * FROM users WHERE idusers = ?`
+        const [userExists] = await connection.query(checkUser, idUser);
+        
+        if (userExists[0].idusers !== req.user.idusers) {
+            return res.status(403).send({
+                message: "Você não tem permissão para criar reviews"
+            })
+        }
+
         const updateByName = `UPDATE book_club_book INNER JOIN books ON book_club_book.book_id = books.idbooks INNER JOIN book_club ON book_club_book.book_club_id = book_club.idbookclub SET book_club_id = (SELECT idbookclub FROM book_club WHERE nomebookclub = ?) WHERE nomebookclub = ?`;
         const clubs = await connection.query(updateByName, [nomebookclub, nomebookclubantigo]);
         connection.release();
@@ -162,6 +224,16 @@ async function deleteClubListByName(req, res) {
     }
 
     try {
+
+        const idUser = req.user.idusers;
+        const checkUser = `SELECT * FROM users WHERE idusers = ?`
+        const [userExists] = await connection.query(checkUser, idUser);
+        
+        if (userExists[0].idusers !== req.user.idusers) {
+            return res.status(403).send({
+                message: "Você não tem permissão para criar reviews"
+            })
+        }
 
         const deleteByNome = `DELETE book_club_book FROM book_club_book INNER JOIN books ON book_club_book.book_id = books.idbooks INNER JOIN book_club ON book_club_book.book_club_id = book_club.idbookclub WHERE nomebookclub = ?`;
         const clubs = await connection.query(deleteByNome, nomebookclub);
